refactor(rpc): implement generated gRPC per-method traits
