+++ crates/rpc/src/server/api/submit_proven_tx.rs
@@ -1,0 +1,137 @@
+use miden_node_proto::generated as proto;
+use miden_node_utils::ErrorReport;
+use miden_node_utils::spawn::spawn_blocking_in_current_span;
+use miden_node_utils::tracing::OpenTelemetrySpanExt;
+use miden_protocol::MIN_PROOF_SECURITY_LEVEL;
+use miden_protocol::transaction::{ProvenTransaction, TxAccountUpdate};
+use miden_protocol::utils::serde::{Deserializable, Serializable};
+use miden_tx::TransactionVerifier;
+use tonic::{Request, Status};
+use tracing::{Span, debug};
+
+use super::{COMPONENT, RpcMode, RpcService, SubmitProvenTxInput, strip_output_note_decorators};
+
+#[tonic::async_trait]
+impl proto::server::rpc_api::SubmitProvenTx for RpcService {
+    type Input = SubmitProvenTxInput;
+    type Output = proto::blockchain::BlockNumber;
+
+    fn decode(request: proto::transaction::ProvenTransaction) -> tonic::Result<Self::Input> {
+        Ok(SubmitProvenTxInput { request, is_authorized_network_tx: false })
+    }
+
+    fn encode(output: Self::Output) -> tonic::Result<proto::blockchain::BlockNumber> {
+        Ok(output)
+    }
+
+    async fn full(
+        &self,
+        request: Request<proto::transaction::ProvenTransaction>,
+    ) -> tonic::Result<proto::blockchain::BlockNumber> {
+        let is_authorized_network_tx = self.is_authorized_network_tx(request.metadata());
+        let mut input = Self::decode(request.into_inner())?;
+        input.is_authorized_network_tx = is_authorized_network_tx;
+        let output = self.handle(input).await?;
+        Self::encode(output)
+    }
+
+    async fn handle(&self, input: Self::Input) -> tonic::Result<Self::Output> {
+        let SubmitProvenTxInput { mut request, is_authorized_network_tx } = input;
+        debug!(target: COMPONENT, ?request);
+
+        let tx = ProvenTransaction::read_from_bytes(&request.transaction).map_err(|err| {
+            Status::invalid_argument(err.as_report_context("invalid transaction"))
+        })?;
+
+        let span = Span::current();
+        span.set_attribute("transaction.id", tx.id());
+        span.set_attribute("account.id", tx.account_id());
+        span.set_attribute("transaction.expires_at", tx.expiration_block_num());
+        span.set_attribute("transaction.reference_block.number", tx.ref_block_num());
+        span.set_attribute("transaction.reference_block.commitment", tx.ref_block_commitment());
+
+        // Verify the reference block is actually part of the chain.
+        self.verify_reference_commitment(tx.ref_block_num(), tx.ref_block_commitment())
+            .await?;
+
+        // Rebuild a new ProvenTransaction with decorators removed from output notes
+        let account_update = TxAccountUpdate::new(
+            tx.account_id(),
+            tx.account_update().initial_state_commitment(),
+            tx.account_update().final_state_commitment(),
+            tx.account_update().account_delta_commitment(),
+            tx.account_update().details().clone(),
+        )
+        .map_err(|e| Status::invalid_argument(e.to_string()))?;
+
+        let stripped_outputs = strip_output_note_decorators(tx.output_notes().iter());
+        let rebuilt_tx = ProvenTransaction::new(
+            account_update,
+            tx.input_notes().iter().cloned(),
+            stripped_outputs,
+            tx.ref_block_num(),
+            tx.ref_block_commitment(),
+            tx.fee(),
+            tx.expiration_block_num(),
+            tx.proof().clone(),
+        )
+        .map_err(|e| Status::invalid_argument(e.to_string()))?;
+        request.transaction = rebuilt_tx.to_bytes();
+
+        // Block post-deployment network-account transactions from user RPC. First-deployment txs
+        // are exempt because the protocol-level allowlist only kicks in once the account exists,
+        // and network accounts must be public, so private-account txs are filtered out up front.
+        //
+        // Skip this check if the client is authorized to send network transactions (ntx-builder).
+        if !is_authorized_network_tx {
+            let candidate_id = (!tx.account_update().initial_state_commitment().is_empty()
+                && tx.account_id().is_public())
+            .then(|| tx.account_id());
+            self.reject_if_any_network_accounts(candidate_id).await?;
+        }
+
+        let tx_id = tx.id();
+        spawn_blocking_in_current_span(move || {
+            TransactionVerifier::new(MIN_PROOF_SECURITY_LEVEL).verify(&tx).map_err(|err| {
+                Status::invalid_argument(format!(
+                    "Invalid proof for transaction {}: {}",
+                    tx_id,
+                    err.as_report()
+                ))
+            })
+        })
+        .await
+        .map_err(|err| {
+            Status::internal(format!("transaction proof verification task failed: {err}"))
+        })??;
+
+        // In full node mode we forward the request to the source.
+        let (block_producer, validator) = match &self.mode {
+            RpcMode::Sequencer { block_producer, validator } => {
+                (block_producer.as_ref(), validator.as_ref())
+            },
+            RpcMode::FullNode { source_rpc } => {
+                return source_rpc
+                    .as_ref()
+                    .clone()
+                    .submit_proven_tx(request)
+                    .await
+                    .map(|response| response.into_inner());
+            },
+        };
+
+        // Transaction inputs must be provided in order to allow for transaction re-execution via
+        // the Validator.
+        if request.transaction_inputs.is_some() {
+            validator.clone().submit_proven_transaction(request.clone()).await?;
+        } else {
+            return Err(Status::invalid_argument("Transaction inputs must be provided"));
+        }
+
+        block_producer
+            .submit_proven_tx(rebuilt_tx)
+            .await
+            .map(Into::into)
+            .map_err(Into::into)
+    }
+}
